+++ Controller/Adminhtml/Image/Chooser.php
@@ -1,18 +1,19 @@
 <?php
+declare(strict_types=1);
 
 namespace MageOS\AdvancedWidget\Controller\Adminhtml\Image;
 
 use Magento\Backend\App\Action\Context;
 use Magento\Cms\Model\Wysiwyg\Images\GetInsertImageContent;
+use Magento\Framework\Controller\Result\Raw;
 use Magento\Framework\Controller\Result\RawFactory;
 use Magento\Framework\Registry;
 use Magento\Framework\UrlInterface;
 use Magento\Store\Model\StoreManagerInterface;
+use Magento\Framework\Exception\NoSuchEntityException;
 
 class Chooser extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images\OnInsert
 {
-    protected $resultRawFactory;
-
     /**
      * @param Context $context
      * @param Registry $coreRegistry
@@ -21,17 +22,20 @@ class Chooser extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images\OnInsert
      * @param StoreManagerInterface $storeManager
      */
     public function __construct(
-        RawFactory $resultRawFactory,
-        private Context $context,
-        private Registry $coreRegistry,
-        private GetInsertImageContent $getInsertImageContent,
-        protected StoreManagerInterface $storeManager,
+        Context $context,
+        Registry $coreRegistry,
+        protected RawFactory $resultRawFactory,
+        protected GetInsertImageContent $getInsertImageContent,
+        protected StoreManagerInterface $storeManager
     ) {
-        $this->resultRawFactory = $resultRawFactory;
         parent::__construct($context, $coreRegistry, $resultRawFactory, $getInsertImageContent);
     }
 
-    public function execute()
+    /**
+     * @return Raw
+     * @throws NoSuchEntityException
+     */
+    public function execute(): Raw
     {
         $data = $this->getRequest()->getParams();
         $mediaUrl = $this->storeManager->getStore()->getBaseUrl(UrlInterface::URL_TYPE_WEB);
@@ -41,8 +45,8 @@ class Chooser extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images\OnInsert
             $data['as_is'],
             isset($data['store']) ? (int) $data['store'] : null
         );
-        $content = str_replace($mediaUrl, '', $content);
-        $content = '/'. ltrim($content, '/');
+        $content = str_replace($mediaUrl, '', (string)$content);
+        $content = '/' . ltrim($content, '/');
         return $this->resultRawFactory->create()->setContents($content);
     }
 }